fix codepen heights
